Refactor Snapshot Area and redirect translation
