+++ app/components/ToolSnapshotArea/SnapshotCallout.js
@@ -1,0 +1,7 @@
+import ContentBlock from 'components/ContentBlock';
+import styled from 'styled-components';
+export default styled(ContentBlock)`
+  font-weight: 800;
+  color: black;
+  margin-bottom: 5px;
+`;
