+++ app/components/ToolSnapshotArea/Viewport.js
@@ -1,0 +1,4 @@
+import styled from 'styled-components';
+export default styled.div`
+  text-align: ${p=>p.lang==='ar'?'right':'left'};
+`;
