+++ app/containers/ToolSnapshotArea/staticText.js
@@ -1,0 +1,14 @@
+export default {
+  viewForm: {
+    id: 'module.form',
+    defaultMessage: 'Go to Form',
+  },
+  seeMore: {
+    id: 'module.bt',
+    defaultMessage: 'See “{{title}}” in Beautiful Trouble'
+  },
+  callOut: {
+    id: 'module.snapshot',
+    defaultMessage: 'Hey, so far this {{type}} is only summarized — but people like you are likely using it in all kinds of ways. Do you have insights to share on how to use this {{type}}? Go ahead and share them through the form below…'
+  }
+}
