+++ app/containers/ToolSnapshotArea/tests/index.test.js
