+++ app/containers/TranslatableStaticText/index.js
@@ -8,10 +8,16 @@ import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
 import { createStructuredSelector } from 'reselect';
 import makeSelectTranslatableStaticText from './selectors';
+import _ from 'lodash';
 
 import { loadLanguage } from './actions';
 
 export class TranslatableStaticText extends React.PureComponent { // eslint-disable-line react/prefer-stateless-function
+
+  constructor() {
+    super();
+    _.templateSettings.interpolate = /{{([\s\S]+?)}}/g;
+  }
 
   //Use to retrieve message from static text
   retrieveMessage() {
@@ -27,11 +33,21 @@ export class TranslatableStaticText extends React.PureComponent { // eslint-disa
     return message || this.props.defaultMessage;
   }
 
+  buildFromTemplate(message) {
+    if (this.props.values) {
+      var compiled = _.template(message);
+      return compiled(this.props.values);
+    }
+
+    return message;
+  }
+
   render() {
     const message = this.retrieveMessage();
+    const extrapolate = this.buildFromTemplate(message);
 
     return (
-      <span>{message}</span>
+      <span>{extrapolate}</span>
     );
   }
 }
