+++ package.json
@@ -221,7 +221,7 @@
     "intl": "1.2.5",
     "invariant": "2.2.2",
     "ip": "1.1.4",
-    "lodash": "4.17.2",
+    "lodash": "^4.17.2",
     "minimist": "1.2.0",
     "react": "15.4.1",
     "react-dom": "15.4.1",
